refactor: move AI analysis files from engine/ to analysis/ feature package
Phase 5 of the clean architecture migration. Groups 20+ analysis files
into coherent sub-packages under analysis/:

- analysis/stream/ — stream parsing, validation, events, progress
- analysis/mcp/ — MCP findings server, dispatch, args
- analysis/prompts/ — prompt building
- analysis/plugins/ — plugin detection, loading, schema validation
- analysis/subagents/ — subagent pool, runner, file queue, JSONL utils
- analysis/runner.py — evaluation orchestrator
- analysis/subprocess.py — AI CLI spawning
- analysis/report.py — report building

Re-exports in old engine/ locations for backward compatibility.
